+++ examples/pipelines/pipeline_example.py
@@ -114,7 +114,7 @@ all_feature_id = config['input_scalar_scaler']['in_features_identifiers'] +\
 # In this example, we aim to predict the ``mach`` field based on two input scalars ``angle_in`` and ``mach_out``, and the mesh node coordinates. To contain memory consumption, we restrict the dataset to the features required for this example:
 
 # %%
-dataset_train = dataset_train.from_features_identifier(all_feature_id)
+dataset_train = dataset_train.extract_dataset_from_identifier(all_feature_id)
 print("dataset_train:", dataset_train)
 print("scalar names =", dataset_train.get_scalar_names())
 print("field names =", dataset_train.get_field_names())